NgVolume and NiiVolume as subclasses of VolumeSource, lazy image and region data loading
